we can get users list from profiles.js
